Add fonts to VariantGrid params and default Variant params
For fonts to be loaded from the templates for single-file test grids,
they must be visible from the grid as a whole, and they must be
aggregated for cases with different fonts declared in different
variants.

This change adds the `fonts` param to grids using `_param_set`, creates
it as an empty list when initializing a default variant.

In a follow-up CL in the stack, a test is created that requires the
loading of fonts and uses single-file variant grids.

Bug: 341213359
Change-Id: Id6d83d26ab4cfa9a98747d573926fc40a24ab565